+++ geopackage-iosTests/extension/schema/columns/GPKGDataColumnsCreateTest.h
@@ -1,0 +1,13 @@
+//
+//  GPKGDataColumnsCreateTest.h
+//  geopackage-iosTests
+//
+//  Created by Brian Osborn on 2/7/24.
+//  Copyright © 2024 NGA. All rights reserved.
+//
+
+#import "GPKGCreateGeoPackageTestCase.h"
+
+@interface GPKGDataColumnsCreateTest : GPKGCreateGeoPackageTestCase
+
+@end