+++ geopackage-iosTests/extension/schema/columns/GPKGDataColumnsImportTest.h
@@ -1,0 +1,13 @@
+//
+//  GPKGDataColumnsImportTest.h
+//  geopackage-iosTests
+//
+//  Created by Brian Osborn on 2/7/24.
+//  Copyright © 2024 NGA. All rights reserved.
+//
+
+#import "GPKGImportGeoPackageTestCase.h"
+
+@interface GPKGDataColumnsImportTest : GPKGImportGeoPackageTestCase
+
+@end